I honestly really love it, could you make it so that in the transactions, there is a Upload Data button as well, where the user can upload a xslx,csv, or pdf file, and when he does, it auto populates the transactions with 5 more dummy rows of data. And remove the Upload Extrato page
The Savings looks great, it would be better if we tracked the users Net Unallocated Cash, which she still has but hasnt put in any bucket or spent from all the past months, and so she can allocate that cash to one of the savings and whenever she adds money to one of the savings, it takes it out of the unallocated cash number

And in the main dashboard, it would be great if we could not only see the net worth over time, which could be this net unallocated cash + savings category type savings or investment category type saving. But also have a way to togle a budget/expenses line chart of the months, where there is a configurable horizontal dotted line showing what the monthly budget is and when she crossed it, which could be configured in the settings.